requirements.txt
pytest
flask
flask-cors
openai
python-dotenv
Pillow
pdf2image
